Fix login state transition error: Rewrite GithubAuth Mutation into class
